+++ RELEASENOTES.md
@@ -99,6 +99,8 @@
         parameter ([#7582](https://github.com/google/ExoPlayer/issues/7582)).
     *   Distinguish between `offsetUs` and `startPositionUs` when passing new
         `SampleStreams` to `Renderers`.
+    *   Fix wrong `MediaPeriodId` for some renderer errors reported by
+        `AnalyticsListener.onPlayerError`.
 *   Video: Pass frame rate hint to `Surface.setFrameRate` on Android R devices.
 *   Track selection:
     *   Add `Player.getTrackSelector`.